fix: 🐛 hover, current page, icons, refactor
Correct hover color for large link pagination. Current page should not
be button/link. New icons. Reuse component styles between button and
link.
