+++ src/Molecules/Pagination/variants/LargeLinks.tsx
@@ -1,44 +1,14 @@
 import React from 'react';
-import styled from 'styled-components';
-import { Flexbox, Link, OldIcon, Typography } from '../../../index';
+import { Flexbox, Typography } from '../../../index';
 import { BrowseLinkProps, PageItemProps, PaginationDefaultLinkProps } from '../Pagination.types';
 import PageItems from '../PageItems';
-import { StyledList, TruncatedPageNumbers } from './Large';
-
-const StyledLink = styled(Link)<{
-  $type: 'chevron' | 'page-item';
-}>`
-  display: flex;
-  height: ${(p) => p.theme.spacing.unit(10)}px;
-  width: ${(p) => p.theme.spacing.unit(10)}px;
-  border: none;
-  padding: 0;
-  justify-content: center;
-  align-items: center;
-  background-color: transparent;
-  ${({ $type, theme }) =>
-    $type === 'chevron' && `box-sizing: border-box; border: 1px solid ${theme.color.inputBorder};`}
-
-  &:hover {
-    background-color: ${(t) => t.theme.color.cta};
-
-    span {
-      color: ${(t) => t.theme.color.buttonText};
-    }
-
-    svg {
-      fill: ${(t) => t.theme.color.buttonText};
-    }
-  }
-`;
-
-const StyledCurrentPage = styled(Flexbox)`
-  background-color: ${(t) => t.theme.color.cta};
-  justify-content: center;
-  align-items: center;
-  height: ${(p) => p.theme.spacing.unit(10)}px;
-  width: ${(p) => p.theme.spacing.unit(10)}px;
-`;
+import {
+  ChevronIcon,
+  StyledCurrentPageBox,
+  StyledList,
+  TruncatedPageNumbers,
+  StyledLink,
+} from './components/LargeComponents';
 
 const PageItem: React.FC<PageItemProps> = ({
   isCurrentPage = false,
@@ -48,22 +18,18 @@ const PageItem: React.FC<PageItemProps> = ({
   href,
 }) => (
   <Flexbox item as="li">
-    {!isCurrentPage ? (
-      <StyledLink $type="page-item" onClick={onClick} aria-label={`${label} ${children}`} to={href}>
-        <Typography
-          type="primary"
-          color={(t) => (isCurrentPage ? t.color.buttonText : t.color.text)}
-          weight="bold"
-        >
-          {children}
-        </Typography>
-      </StyledLink>
-    ) : (
-      <StyledCurrentPage container>
+    {isCurrentPage ? (
+      <StyledCurrentPageBox container>
         <Typography type="primary" color={(t) => t.color.buttonText} weight="bold">
           {children}
         </Typography>
-      </StyledCurrentPage>
+      </StyledCurrentPageBox>
+    ) : (
+      <StyledLink $type="page-item" onClick={onClick} aria-label={`${label} ${children}`} to={href}>
+        <Typography type="primary" color={(t) => t.color.text} weight="bold">
+          {children}
+        </Typography>
+      </StyledLink>
     )}
   </Flexbox>
 );
@@ -71,11 +37,7 @@ const PageItem: React.FC<PageItemProps> = ({
 const ChevronLink: React.FC<BrowseLinkProps> = ({ direction, onClick, label, href }) => (
   <Flexbox item container alignItems="center">
     <StyledLink $type="chevron" onClick={onClick} aria-label={label} to={href}>
-      {direction === 'left' ? (
-        <OldIcon.ChevronLeft inline size={3} />
-      ) : (
-        <OldIcon.ChevronRight inline size={3} />
-      )}
+      <ChevronIcon direction={direction} />
     </StyledLink>
   </Flexbox>
 );
