+++ src/Molecules/Pagination/variants/Regular.tsx
@@ -1,27 +1,9 @@
 import React from 'react';
-import styled from 'styled-components';
-import { Button, Flexbox, OldIcon, Typography } from '../../../index';
+import { Button, Flexbox, Typography } from '../../../index';
 import List from '../../../Atoms/List';
 import PageItems from '../PageItems';
 import { BrowseButtonProps, PageItemProps, PaginationDefaultProps } from '../Pagination.types';
-
-const StyledFlexbox = styled(Flexbox)<{ $numberOfPages: number }>`
-  width: ${(p) => {
-    const NUMBER_WIDTH = p.theme.spacing.unit(4.5);
-    const CHEVRON_PADDING = 2 * p.theme.spacing.unit(2);
-    const BETWEEN_8_AND_13_PAGES_WIDTH = p.theme.spacing.unit(38);
-    const BETWEEN_14_AND_101_PAGES_WIDTH = p.theme.spacing.unit(42);
-    const MORE_THAN_101_PAGES_WIDTH = p.theme.spacing.unit(50);
-
-    if (p.$numberOfPages < 8) return p.$numberOfPages * NUMBER_WIDTH + CHEVRON_PADDING;
-
-    if (p.$numberOfPages >= 8 && p.$numberOfPages <= 13) return BETWEEN_8_AND_13_PAGES_WIDTH;
-
-    if (p.$numberOfPages >= 14 && p.$numberOfPages <= 101) return BETWEEN_14_AND_101_PAGES_WIDTH;
-
-    return MORE_THAN_101_PAGES_WIDTH;
-  }}px;
-`;
+import { ChevronIcon, StyledFlexbox, TruncatedPageNumbers } from './components/RegularComponents';
 
 const PageNumberItem: React.FC<PageItemProps> = ({
   isCurrentPage = false,
@@ -30,34 +12,23 @@ const PageNumberItem: React.FC<PageItemProps> = ({
   label,
 }) => (
   <Flexbox item as="li">
-    <Button
-      variant="neutral"
-      onClick={!isCurrentPage ? onClick : undefined}
-      aria-label={`${label} ${children}`}
-      aria-disabled={isCurrentPage}
-    >
-      <Typography
-        type="secondary"
-        color={(t) => (isCurrentPage ? t.color.cta : t.color.text)}
-        weight="bold"
-      >
+    {isCurrentPage ? (
+      <Typography type="secondary" color={(t) => t.color.cta} weight="bold">
         {children}
       </Typography>
-    </Button>
-  </Flexbox>
-);
-
-const TruncatedPageNumbers = () => (
-  <Flexbox item container alignItems="center" as="li">
-    <Typography type="secondary" weight="bold">
-      ...
-    </Typography>
+    ) : (
+      <Button variant="neutral" onClick={onClick} aria-label={`${label} ${children}`}>
+        <Typography type="secondary" color={(t) => t.color.text} weight="bold">
+          {children}
+        </Typography>
+      </Button>
+    )}
   </Flexbox>
 );
 
 const ChevronButton: React.FC<BrowseButtonProps> = ({ direction = 'left', onClick, label }) => (
   <Button variant="neutral" onClick={onClick} aria-label={label}>
-    {direction === 'left' ? <OldIcon.ChevronLeft size={2} /> : <OldIcon.ChevronRight size={2} />}
+    <ChevronIcon direction={direction} />
   </Button>
 );
 
@@ -100,4 +71,3 @@ const Regular: React.FC<PaginationDefaultProps> = ({
 };
 
 export default React.memo(Regular);
-export { StyledFlexbox, TruncatedPageNumbers };
